+++ lib/src/objects.dart
@@ -1,7 +1,10 @@
 import 'dart:convert';
+export 'constants.dart';
 export 'objects/exchange_info.dart';
+export 'objects/acc_objects.dart';
+export 'objects/orders_objects.dart';
 
-enum KLinesType { spot, futures, futuresCoin }
+// enum KLinesType { spot, futures, futuresCoin }
 
 enum HttpMethod { get, post, put, delete }
 
@@ -34,7 +37,7 @@ class BinanceApiException implements Exception {
   }
 
   @override
-  String toString() => message;
+  String toString() => '$code $message';
 }
 
 class SymbolProduct {
@@ -78,5 +81,53 @@ class SymbolProduct {
   @override
   String toString() {
     return '$baseAsset:$quoteAsset $status';
+  }
+}
+
+class CandleStick {
+  /// Kline/candlestick
+  final DateTime openTime; // Kline open time
+  final double openPrice; // Open price
+  final double highPrice; // High price
+  final double lowPrice; // Low price
+  final double closePrice; // Close price
+  final double volume; // Volume
+  final DateTime closeTime; // Kline Close time
+  final double quoteVolume; // Quote asset volume
+  final int numberOfTrades; // Number of trades
+  final double takerBaseVolume; // Taker buy base asset volume
+  final double takerQuoteVolume; // Taker buy quote asset volume
+
+  CandleStick(List m)
+      : openTime = DateTime.fromMillisecondsSinceEpoch(m[0]),
+        openPrice = double.parse(m[1]),
+        highPrice = double.parse(m[2]),
+        lowPrice = double.parse(m[3]),
+        closePrice = double.parse(m[4]),
+        volume = double.parse(m[5]),
+        closeTime = DateTime.fromMillisecondsSinceEpoch(m[6]),
+        quoteVolume = double.parse(m[7]),
+        numberOfTrades = m[8],
+        takerBaseVolume = double.parse(m[9]),
+        takerQuoteVolume = double.parse(m[10]);
+
+  @override
+  String toString() {
+    return 'close: $closePrice volume: $volume';
+  }
+}
+
+class AvgPrice {
+  /// AvgPrice
+  final int mins;
+  final double price;
+
+  AvgPrice(Map m)
+      : mins = m['mins'],
+        price = double.parse(m['price']);
+
+  @override
+  String toString() {
+    return 'mins: $mins price: $price';
   }
 }